CRYP-49: Added A New Endpoint To Get List Of Offers Made By The Artist

+++ src/CryptoAxus.Common/Response/PaginationData.cs
@@ -2,13 +2,13 @@
 
 public class PaginationData
 {
-    public int? TotalRecords { get; set; }
+    public long? TotalRecords { get; set; }
 
     public int? CurrentPage { get; set; }
 
     public int? PageSize { get; set; }
 
-    public int? TotalPages { get; set; }
+    public long? TotalPages { get; set; }
 
     public bool? HasPrevious { get; set; }
 
@@ -20,31 +20,31 @@ public class PaginationData
 
     public PaginationData() { }
 
-    public PaginationData(int totalRecords, int pageSize)
+    public PaginationData(long totalRecords, int pageSize)
     {
         TotalRecords = totalRecords;
         PageSize = pageSize;
-        TotalPages = (int)Math.Ceiling(totalRecords / (double)pageSize);
+        TotalPages = (long)Math.Ceiling(totalRecords / (double)pageSize);
     }
 
-    public PaginationData(int? totalRecords, int? pageNumber, int? pageSize)
+    public PaginationData(long? totalRecords, int? pageNumber, int? pageSize)
     {
         TotalRecords = totalRecords;
         CurrentPage = pageNumber;
         PageSize = pageSize;
         if (totalRecords is not null && pageSize is not null)
-            TotalPages = (int)Math.Ceiling(totalRecords.Value / (double)pageSize.Value);
+            TotalPages = (long)Math.Ceiling(totalRecords.Value / (double)pageSize.Value);
         HasPrevious = CurrentPage > 1;
         HasNext = CurrentPage < TotalPages;
     }
 
-    public PaginationData(int? totalRecords, int? pageNumber, int? pageSize, string url)
+    public PaginationData(long? totalRecords, int? pageNumber, int? pageSize, string url)
     {
         TotalRecords = totalRecords;
         CurrentPage = pageNumber;
         PageSize = pageSize;
         if (totalRecords is not null && pageSize is not null)
-            TotalPages = (int)Math.Ceiling(totalRecords.Value / (double)pageSize.Value);
+            TotalPages = (long)Math.Ceiling(totalRecords.Value / (double)pageSize.Value);
         HasPrevious = CurrentPage > 1;
         HasNext = CurrentPage < TotalPages;
         PreviousPageLink = pageNumber > 1 ? $"{url}?pageNumber={pageNumber - 1}&pageSize={pageSize}" : null;
